+++ .replit
@@ -1,7 +1,7 @@
-run = "python scraper.py"
+run = "python main.py"
 modules = ["python-3.10"]
 [nix]
 channel = "stable-23_11"
 
 [deployment]
-run = ["sh", "-c", "python scraper.py"]
+run = ["sh", "-c", "python main.py"]
