--- Chattanooga_TN_United_States_05_21_2024_To_05_22_2024.csv
@@ -1,27 +1,0 @@
-Location,Check-in Date,Check-out Date,Hotel Name,Room Type,Price (INR)
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Hotel Indigo: Chattanooga - Downtown,1 King Bed Standard,"20,156"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Hotel Indigo: Chattanooga - Downtown,Standard Room,"20,156"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Hotel Indigo: Chattanooga - Downtown,1 King Bed 1 Bedroom Suite,"24,203"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Staybridge Suites: Chattanooga Dwtn - Conv Ctnr,2 Double Suite Sofa Bed,"19,344"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Staybridge Suites: Chattanooga Dwtn - Conv Ctnr,Standard Suite,"19,344"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Chattanooga-Lookout Mtn,Standard Room,"16,569"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Chattanooga-Lookout Mtn,1 King Bed Suite Wet Bar Sofa Bed,"17,362"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Chattanooga-Lookout Mtn,1 King Suite Sofa Bed,"17,362"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Chattanooga-Lookout Mtn,Specialty Room,"18,154"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Chattanooga-Lookout Mtn,1 King Bed Standard Corner Room,"18,947"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,avid hotels: Chattanooga South - Ringgold,1 King Standard,"19,740"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,avid hotels: Chattanooga South - Ringgold,2 Queen Standard,"19,740"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,avid hotels: Chattanooga South - Ringgold,Standard Room,"19,740"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Kimball,2 Queen Standard,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Kimball,1 King Suite Additional Living Area,"17,362"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Kimball,Standard Room,"17,362"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Kimball,2 Queen Suite Sofa Bed,"18,947"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn: Cleveland,1 King Bed Standard With Trundle Bed,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn: Cleveland,1 King Bed Standard Sofa Bed,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn: Cleveland,2 Queen Beds Standard,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn: Cleveland,STANDARD ROOM,"16,173"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn: Cleveland,1 King Suite With Additional Living Area,"16,965"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Cleveland Northwest,1 King Bed Standard,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Cleveland Northwest,2 Queen Beds Standard,"15,776"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Cleveland Northwest,Standard Room,"18,947"
-"Chattanooga, TN, United States",05/21/2024,05/22/2024,Holiday Inn Express & Suites: Cleveland Northwest,2 Queen Beds 1 Bedroom Suite,"20,136"
