+++ info.md
@@ -1,0 +1,6 @@
+1. Chattanooga, TN, United States => 18
+2. Chattanooga Lovell Airport, TN, United States => 19
+3. Louisville, KY, United States => 21
+4. College Station, TX, United States => 7
+5. Baton Rouge, LA, United States LA => 12
+6. Auburn, Alabama, United States => 7
