+++ input_data.py
@@ -1,0 +1,6 @@
+from datetime import datetime
+
+
+destinations = ['College Station, TX, United States', 'Auburn, Alabama, United States']
+check_in_date = datetime(2024, 5, 23).strftime('%m/%d/%Y')
+check_out_date = datetime(2024, 5, 24).strftime('%m/%d/%Y')
